requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.15
sentence-transformers==3.1.1
openai==1.51.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
